+++ designer/client/cypress/e2e/components.cy.ts
@@ -101,7 +101,7 @@ describe("Components list", () => {
         cy.contains(/^usages$/i).click();
         cy.contains(/^≥ 1$/i).click();
         cy.get("body").click();
-        cy.get("[data-id=filter] [data-testid=LinkIcon]").click();
+        cy.get("[data-id=builtin-filter] [data-testid=LinkIcon]").click();
         cy.contains(/^status$/i).click();
         cy.contains(/^running$/i).click();
         cy.contains(/^no rows$/i).should("be.visible");
@@ -166,7 +166,7 @@ describe("Components list", () => {
     it("should filter usages", () => {
         cy.createTestProcess(`${seed}_xxx`, "testProcess2");
 
-        cy.visit("/components/usages/filter");
+        cy.visit("/components/usages/builtin-filter");
 
         cy.get("input[type=text]").type("8 xxx");
         cy.matchQuery("?TEXT=8+xxx");
@@ -196,7 +196,7 @@ describe("Components list", () => {
         cy.contains(/^ok$/i).click();
 
         cy.viewport(1400, 600);
-        cy.visit("/components/usages/filter");
+        cy.visit("/components/usages/builtin-filter");
 
         cy.contains(/^other$/i).click();
         cy.get("[role=menu]").find("li[role=menuitem]").as("options");