+++ designer/server/src/main/scala/db/migration/hsql/V1_047__RegenerateComponentsUsages.scala
@@ -3,6 +3,6 @@ package db.migration.hsql
 import db.migration.InvalidateComponentsUsagesCache
 import slick.jdbc.{HsqldbProfile, JdbcProfile}
 
-class V1_039__FillComponentsUsages extends InvalidateComponentsUsagesCache {
+class V1_047__RegenerateComponentsUsages extends InvalidateComponentsUsagesCache {
   override protected lazy val profile: JdbcProfile = HsqldbProfile
 }